+++ CHANGES.md
@@ -33,6 +33,7 @@ Changes in 12.8
 - Security fix arbitrary Table Row Deletion via `$_REQUEST['table']` in Medical.inc.php, thanks to @lighthousekeeper1212
 - SQL performance: use explicit JOINs & remove School column in Portal.php
 - SQL performance + harmonize SQL logic for Block Periods in AddAbsences.phph, Administration.php, TakeAttendance.php, TeacherCompletion.php, UpdateAttendanceDaily.fnc.php & Courses.fnc.php
+- Skip CSP violation triggered by Facebook Click Identifier in SaveReport.php
 
 Changes in 12.7.4
 -----------------